Adiciona Jest com teste de contagem de visitas e ajuste para ambiente de teste
