+++ package.json
@@ -14,6 +14,7 @@
     "redis": "^4.7.0"
   },
   "devDependencies": {
-    "jest": "^29.7.0"
+    "jest": "^29.7.0",
+    "supertest": "^7.0.0"
   }
 }
